bug on difficult level

+++ Assets/Scripts/GameVariables.cs
@@ -8,6 +8,6 @@ public class GameVariables : MonoBehaviour
     public static int NumberOfQuestions;
     public static QuestionBank QuizQuestionBank;
     public static bool PlaySounds = true;
-    public static int DifficultyLevel = 0;
+    public static float DifficultyLevel = 0;
     public static string SettingsNumberOfQuestions = "5";
 }
